+++ src/main/java/com/jb/CouponSystemProjectP2/Controllers/CompanyController.java
@@ -2,9 +2,11 @@ package com.jb.CouponSystemProjectP2.Controllers;
 
 import com.jb.CouponSystemProjectP2.Beans.Category;
 import com.jb.CouponSystemProjectP2.Beans.Coupon;
+import com.jb.CouponSystemProjectP2.Beans.UserType;
 import com.jb.CouponSystemProjectP2.Exceptions.CompanyException;
 import com.jb.CouponSystemProjectP2.Exceptions.CouponNotFoundException;
 import com.jb.CouponSystemProjectP2.Exceptions.TokenException;
+import com.jb.CouponSystemProjectP2.Exceptions.UnauthorizedUserException;
 import com.jb.CouponSystemProjectP2.Security.JWTutil;
 import com.jb.CouponSystemProjectP2.Services.CompanyService;
 import io.jsonwebtoken.ExpiredJwtException;
@@ -24,7 +26,10 @@ public class CompanyController {
 
     // -------------------CREATE--------------------
     @PostMapping("/coupons/add") // http://localhost:8080/company/coupons/add
-    public ResponseEntity<?> createNewCoupon(@RequestHeader(name = "Authorization") String token, @RequestBody Coupon coupon) throws CompanyException, TokenException {
+    public ResponseEntity<?> createNewCoupon(@RequestHeader(name = "Authorization") String token, @RequestBody Coupon coupon) throws CompanyException, TokenException, UnauthorizedUserException {
+        if (!jwtUtil.getUserTypeFromToken(token).equals(UserType.COMPANY)){
+            throw new UnauthorizedUserException();
+        }
         try {
             if (jwtUtil.isTokenValid(token)) {
                 companyService.createCoupon(jwtUtil.getIdFromToken(token), coupon);
@@ -41,14 +46,20 @@ public class CompanyController {
 
     //---------------READ----------------------
     @GetMapping("/coupons/all") // http://localhost:8080/company/coupons/all
-    public ResponseEntity<?> getAllCompanyCoupons(@RequestHeader(name = "Authorization") String token) throws CouponNotFoundException {
+    public ResponseEntity<?> getAllCompanyCoupons(@RequestHeader(name = "Authorization") String token) throws CouponNotFoundException, UnauthorizedUserException {
+        if (!jwtUtil.getUserTypeFromToken(token).equals(UserType.COMPANY)){
+            throw new UnauthorizedUserException();
+        }
         return ResponseEntity.ok()
                 .header("Authorization", jwtUtil.generateToken(token))
                 .body(companyService.readAllCompanyCoupons(jwtUtil.getIdFromToken(token)));
     }
 
     @GetMapping("/coupons/cat/{category}") // http://localhost:8080/company/coupons/cat/{category}
-    public ResponseEntity<?> getCompanyCouponsByCategory(@RequestHeader(name = "Authorization") String token, @PathVariable Category category) throws CouponNotFoundException {
+    public ResponseEntity<?> getCompanyCouponsByCategory(@RequestHeader(name = "Authorization") String token, @PathVariable Category category) throws CouponNotFoundException, UnauthorizedUserException {
+        if (!jwtUtil.getUserTypeFromToken(token).equals(UserType.COMPANY)){
+            throw new UnauthorizedUserException();
+        }
         return ResponseEntity.ok()
                 .header("Authorization", jwtUtil.generateToken(token))
                 .body(companyService.readCompanyCouponsByCategory(jwtUtil.getIdFromToken(token), category));
@@ -56,14 +67,20 @@ public class CompanyController {
 
 
     @GetMapping("/coupons/max/{price}") // http://localhost:8080/company/coupons/max/{price}
-    public ResponseEntity<?> getCompanyCouponsByMaxPrice(@RequestHeader(name = "Authorization") String token, @PathVariable double price) throws CouponNotFoundException {
+    public ResponseEntity<?> getCompanyCouponsByMaxPrice(@RequestHeader(name = "Authorization") String token, @PathVariable double price) throws CouponNotFoundException, UnauthorizedUserException {
+        if (!jwtUtil.getUserTypeFromToken(token).equals(UserType.COMPANY)){
+            throw new UnauthorizedUserException();
+        }
         return ResponseEntity.ok()
                 .header("Authorization", jwtUtil.generateToken(token))
                 .body(companyService.readCompanyCouponsByMaxPrice(jwtUtil.getIdFromToken(token), price));
     }
 
     @GetMapping("/details") // http://localhost:8080/company/details
-    public ResponseEntity<?> getCompanyDetail(@RequestHeader(name = "Authorization") String token) {
+    public ResponseEntity<?> getCompanyDetail(@RequestHeader(name = "Authorization") String token) throws UnauthorizedUserException {
+        if (!jwtUtil.getUserTypeFromToken(token).equals(UserType.COMPANY)){
+            throw new UnauthorizedUserException();
+        }
         return ResponseEntity.ok()
                 .header("Authorization", jwtUtil.generateToken(token))
                 .body(companyService.readCompanyDetails(jwtUtil.getIdFromToken(token)));
@@ -71,7 +88,10 @@ public class CompanyController {
 
     //-------------------------UPDATE------------------
     @PutMapping("/coupons/update") // http://localhost:8080/company/coupons/update
-    public ResponseEntity<?> updateCoupon(@RequestHeader(name = "Authorization") String token, @RequestBody Coupon coupon) throws CouponNotFoundException {
+    public ResponseEntity<?> updateCoupon(@RequestHeader(name = "Authorization") String token, @RequestBody Coupon coupon) throws CouponNotFoundException, UnauthorizedUserException {
+        if (!jwtUtil.getUserTypeFromToken(token).equals(UserType.COMPANY)){
+            throw new UnauthorizedUserException();
+        }
         companyService.updateCoupon(jwtUtil.getIdFromToken(token), coupon);
         return ResponseEntity.accepted()
                 .header("Authorization", jwtUtil.generateToken(token))
@@ -80,7 +100,10 @@ public class CompanyController {
 
     // ----------------------DELETE---------------------
     @DeleteMapping("/coupons/delete/{id}") // http://localhost:8080/company/coupons/delete/{id}
-    public ResponseEntity<?> deleteCouponById(@RequestHeader(name = "Authorization") String token, @PathVariable int couponId) throws CouponNotFoundException {
+    public ResponseEntity<?> deleteCouponById(@RequestHeader(name = "Authorization") String token, @PathVariable int couponId) throws CouponNotFoundException, UnauthorizedUserException {
+        if (!jwtUtil.getUserTypeFromToken(token).equals(UserType.COMPANY)){
+            throw new UnauthorizedUserException();
+        }
         companyService.deleteCouponById(couponId);
         return ResponseEntity.accepted()
                 .header("Authorization", jwtUtil.generateToken(token))
